modify#6feat(login, register): 폴더 분리
