+++ frontEnd/public/login/view.js
@@ -1,0 +1,12 @@
+import { getLoginTemplate } from "./template";
+
+const loginView = () => { 
+    const $loginContainer = document.createElement("section");
+    $loginContainer.classList.add("loginArea");
+    
+    $loginContainer.innerHTML = getLoginTemplate();
+
+    return $loginContainer;
+}
+
+export { loginView }
