+++ frontEnd/public/register/view.js
@@ -1,0 +1,12 @@
+import { getRegisterTemplate } from "./template";
+
+const registerView = () => {
+    const $registerContainer = document.createElement("section");
+    $registerContainer.classList.add("main");
+
+    $registerContainer.innerHTML = getRegisterTemplate();
+
+    return $registerContainer;
+}
+
+export { registerView }
