+++ frontEnd/public/temp/helper/helperFunction.js
